pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tgclaude"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "claude-agent-sdk>=0.0.13",
    "markdown-it-py>=3.0",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "python-dotenv>=1.0",
    "zoneinfo; python_version < '3.9'",
]

[project.scripts]
tgclaude = "tgclaude.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
